fix: menu item clickable area
