+++ src/components/card/card.tsx
@@ -3,27 +3,7 @@ import tw from 'twin.macro';
 
 import { cardBg } from '$/styles/common';
 
-type PropsOf<E extends keyof JSX.IntrinsicElements | React.JSXElementConstructor<any>> =
-  JSX.LibraryManagedAttributes<E, React.ComponentPropsWithRef<E>>;
-
-export interface BoxOwnProps<E extends React.ElementType = React.ElementType> {
-  as?: E;
-}
-
-export type BoxProps<E extends React.ElementType> = BoxOwnProps<E> &
-  Omit<PropsOf<E>, keyof BoxOwnProps>;
-
-export type PolymorphicComponentProps<E extends React.ElementType, P> = P & BoxProps<E>;
-
-const defaultElement = 'div';
-
-export const Box = React.forwardRef(function Box(
-  { as, ...restProps }: BoxOwnProps,
-  ref: React.Ref<Element>,
-) {
-  const Element = as || defaultElement;
-  return <Element ref={ref} {...restProps} />;
-}) as <E extends React.ElementType = typeof defaultElement>(props: BoxProps<E>) => JSX.Element;
+import { PolymorphicComponentProps, Box } from '../box';
 
 type CustomComponentOwnProps = {
   /**
